requirements.txt
attrs==20.2.0
Authlib==0.15
boto3==1.16.9
botocore==1.19.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chardet2==2.0.3
click==7.1.2
cryptography==3.1.1
flash==1.0.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-Moment==0.10.0
Flask-OAuth2-Login==0.0.9
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pycryptodome==3.9.8
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
secret==0.8
six==1.15.0
SQLAlchemy==1.3.19
tabulate==0.8.7
toml==0.10.2
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
